add argument handler and reverse page order for right pages

+++ pdfmkduplex.sh
@@ -1,5 +1,37 @@
 #!/bin/bash
 
+while getopts ":l:r:d:h" opt; do
+  case ${opt} in
+    l )
+      left_pdf=$OPTARG
+      ;;
+    r )
+      right_pdf=$OPTARG
+      ;;
+    d )
+      destination=$OPTARG
+      ;;
+    h )
+      echo "usage: pdfmkduplex.sh -l <left.pdf> -r <right.pdf> -d <destination>"
+      exit 0
+      ;;
+    \? )
+      echo "Invalid option: $OPTARG" 1>&2
+      exit 1
+      ;;
+    : )
+      echo "Invalid option: $OPTARG requires an argument" 1>&2
+      exit 1
+      ;;
+  esac
+done
+shift $((OPTIND -1))
+
+if [[ -z "$left_pdf" || -z "$right_pdf" || -z "$destination" ]]; then
+  echo "usage: pdfmkduplex.sh -l <left.pdf> -r <right.pdf> -d <destination>"
+  exit 1
+fi
+
 if [[ $# -lt 3 ]]; then
     echo "error: 3 arguments expected, $# where given"
     echo "usage pdfmkduplex.sh <left.pdf> <right.pdf> <destination>"
@@ -19,6 +51,13 @@ if [ ! -d $(dirname $3) ]; then
     exit 1
 fi
 
+lenl=`pdfinfo $1 2> /dev/null | awk '/^Pages:/ {print $2}'`
+lenr=`pdfinfo $2 2> /dev/null | awk '/^Pages:/ {print $2}'`
+if [[ ${lenl} -ne ${lenr} ]]; then
+    echo "error: length of documents left ($lenl) and of right ($lenr) not equal"
+    exit 1
+fi
+
 rm -rf /tmp/mkduplex-*
 TMPDIRNAME=/tmp/mkduplex-$$
 mkdir -p ${TMPDIRNAME}
@@ -30,7 +69,8 @@ pdfseparate -f $count -l $count $1 ${TMPDIRNAME}/${count0}_l.pdf
 RETCODE=$?
 while [[ $RETCODE -eq 0 ]]
 do
-    pdfseparate -f $count -l $count $2 ${TMPDIRNAME}/${count0}_r.pdf
+    countinv=$[ $lenr + 1 - $count ]
+    pdfseparate -f $countinv -l $countinv $2 ${TMPDIRNAME}/${count0}_r.pdf
     if [[ $? -ne 0 ]]
     then
 	echo "$2 has less page than $1 ?"
